# ai-agent-fe/openapi.yaml
openapi: 3.1.0
info:
  title: AI Agent API
  description: API for the AI Agent frontend and service interaction
  version: 0.1.0
servers:
  - url: /api
    description: Development server

# --- Components Definition (Single top-level key) ---
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: A description of the error
      required:
        - error

    # --- User Schemas ---
...
    AdkRequestPayload:
      type: object
      required:
        - user_id
        - session_id
        - new_message
      properties:
        user_id:
          type: string
          description: User ID for the conversation
        session_id:
          type: string
          description: Session ID for the conversation
        new_message:
          type: object
          properties:
            role:
              type: string
              enum: [user]
            text:
              type: string
              description: The message text
          required:
            - role
            - text
        context:
          type: object
          description: Additional context for the conversation
          additionalProperties: true
    AdkResponsePayload:
      type: object
      properties:
        content:
          type: object
          properties:
            parts:
              type: array
              items:
                type: object
                properties:
                  text:
                    type: string
                  mime_type:
                    type: string
            next_actions:
              type: array
              items:
                type: object
        events:
          type: array
          items:
            type: object
            properties:
              event_type:
                type: string
              content:
                type: object
                additionalProperties: true

    # --- Project Schemas ---
...
    DownloadResponse:
        type: object
        properties:
            downloadURL:
                type: string
                format: url
                nullable: true
            error:
                type: string
                nullable: true
            message:
                type: string
                nullable: true

# --- Reusable Responses Definition ---
...
paths:
  /adk:
    post:
      tags: [ADK]
      summary: Proxy to the ADK agent
      description: Forwards user messages to the ADK agent and returns the agent's response
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdkRequestPayload'
      responses:
        '200':
          description: ADK agent response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdkResponsePayload'
        '400':
          description: Bad Request (e.g., missing required fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /projects/{projectId}/download:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
        description: ID of the project to download.
    get:
      tags: [Projects]
      summary: Get project download link
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Download link generated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Project not found or no downloadable content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

# --- Tags Definition (Keep tags at the top level) ---
...